feat(frontend): evaluation integration for new apps
