+++ agenta-web/src/components/EvaluationTable/components/ParamsFormWithRun.tsx
@@ -1,0 +1,69 @@
+import {CaretRightOutlined} from "@ant-design/icons"
+import {Button, Form} from "antd"
+
+import {useLegacyVariants} from "@/lib/hooks/useLegacyVariant"
+
+import ParamsForm from "../../OldPlayground/ParamsForm/ParamsForm"
+import {useSingleModelEvaluationTableStyles} from "../assets/styles"
+
+import type {Evaluation} from "@/lib/Types"
+import type {SingleModelEvaluationRow} from "../types"
+
+/**
+ *
+ * @param evaluation - Evaluation object
+ * @param evaluationScenarios - Evaluation rows
+ * @param columnsCount - Number of variants to compare face to face (per default 2)
+ * @returns
+ */
+const ParamsFormWithRun = ({
+    evaluation,
+    record,
+    rowIndex,
+    onRun,
+    onParamChange,
+    variantData,
+}: {
+    record: SingleModelEvaluationRow
+    rowIndex: number
+    evaluation: Evaluation
+    onRun: () => void
+    onParamChange: (name: string, value: any) => void
+    variantData: ReturnType<typeof useLegacyVariants>
+}) => {
+    const classes = useSingleModelEvaluationTableStyles()
+    const [form] = Form.useForm()
+
+    return (
+        <div>
+            {evaluation.testset.testsetChatColumn ? (
+                evaluation.testset.csvdata[rowIndex][evaluation.testset.testsetChatColumn] || " - "
+            ) : (
+                <ParamsForm
+                    isChatVariant={false}
+                    onParamChange={onParamChange}
+                    inputParams={
+                        variantData[0].inputParams?.map((item) => ({
+                            ...item,
+                            value: record.inputs.find((ip) => ip.input_name === item.name)
+                                ?.input_value,
+                        })) || []
+                    }
+                    onFinish={onRun}
+                    form={form}
+                />
+            )}
+
+            <div className={classes.inputTestBtn}>
+                <Button
+                    onClick={evaluation.testset.testsetChatColumn ? onRun : form.submit}
+                    icon={<CaretRightOutlined />}
+                >
+                    Run
+                </Button>
+            </div>
+        </div>
+    )
+}
+
+export default ParamsFormWithRun
