+++ agenta-web/src/components/Evaluations/EvaluationCardView/VariantAlphabet.tsx
@@ -1,7 +1,7 @@
 import {Typography} from "antd"
 import React from "react"
 import {createUseStyles} from "react-jss"
-import {VARIANT_COLORS} from "."
+import {VARIANT_COLORS} from "./assets/styles"
 
 type StyleProps = {
     color: string
