+++ agenta-web/src/components/HumanEvaluationModal/types.d.ts
@@ -1,0 +1,5 @@
+export interface HumanEvaluationModalProps {
+    isEvalModalOpen: boolean
+    setIsEvalModalOpen: React.Dispatch<React.SetStateAction<boolean>>
+    evaluationType: "single_model_test" | "human_a_b_testing"
+}
